# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "cell-analysis-backend",
        "description": "Backend API for cancer cell analysis application, apoptosis and viable cell",
        "contact": {
            "name": "Your Name",
            "email": "your.email@example.com"
        },
        "license": {
            "name": ""
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/images/{image_id}": {
            "get": {
                "tags": [
                    "Image Management"
                ],
                "summary": "Get details of a specific image",
                "operationId": "get_image",
                "parameters": [
                    {
                        "name": "image_id",
                        "in": "path",
                        "description": "Image ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_ImageDetailResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Image not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Image Management"
                ],
                "summary": "Delete an image (soft delete)",
                "operationId": "delete_image",
                "parameters": [
                    {
                        "name": "image_id",
                        "in": "path",
                        "description": "Image ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_DeleteImageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Image not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Image Management"
                ],
                "summary": "Rename an image",
                "operationId": "rename_image",
                "parameters": [
                    {
                        "name": "image_id",
                        "in": "path",
                        "description": "Image ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RenameImageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Image renamed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse_ImageResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid filename"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Image not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AnalysisHistoryItem": {
                "type": "object",
                "description": "Analysis history item for image detail",
                "required": [
                    "job_id",
                    "status"
                ],
                "properties": {
                    "ai_model_version": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "finished_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "job_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "type": "string"
                    }
                }
            },
...
            "ApiError": {
                "type": "object",
                "description": "API error structure",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "ApiResponse_DeleteImageResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "Delete image response",
                        "required": [
                            "message"
                        ],
                        "properties": {
                            "message": {
                                "type": "string"
                            }
                        }
                    },
                    "error": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiError"
                            }
                        ]
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "ApiResponse_ImageDetailResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "Image detail response (with analysis history)",
                        "required": [
                            "image_id",
                            "folder_id",
                            "original_filename",
                            "file_url",
                            "file_size",
                            "mime_type",
                            "analysis_history",
                            "uploaded_at"
                        ],
                        "properties": {
                            "analysis_history": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/AnalysisHistoryItem"
                                }
                            },
                            "file_size": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "file_url": {
                                "type": "string"
                            },
                            "folder_id": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "image_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "metadata": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ImageMetadataResponse"
                                    }
                                ]
                            },
                            "mime_type": {
                                "type": "string"
                            },
                            "original_filename": {
                                "type": "string"
                            },
                            "uploaded_at": {
                                "type": "string"
                            }
                        }
                    },
                    "error": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiError"
                            }
                        ]
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "ApiResponse_ImageResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "description": "Single image response",
                        "required": [
                            "image_id",
                            "folder_id",
                            "original_filename",
                            "file_size",
                            "mime_type",
                            "has_analysis",
                            "uploaded_at"
                        ],
                        "properties": {
                            "file_size": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "folder_id": {
                                "type": "integer",
                                "format": "int32"
                            },
                            "has_analysis": {
                                "type": "boolean"
                            },
                            "image_id": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "metadata": {
                                "oneOf": [
                                    {
                                        "type": "null"
                                    },
                                    {
                                        "$ref": "#/components/schemas/ImageMetadataResponse"
                                    }
                                ]
                            },
                            "mime_type": {
                                "type": "string"
                            },
                            "original_filename": {
                                "type": "string"
                            },
                            "uploaded_at": {
                                "type": "string"
                            }
                        }
                    },
                    "error": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ApiError"
                            }
                        ]
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "ImageMetadataResponse": {
                "type": "object",
                "description": "Image metadata",
                "properties": {
                    "height": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int32",
                        "minimum": 0
                    },
                    "width": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int32",
                        "minimum": 0
                    }
                }
            },
...
            "RenameImageRequest": {
                "type": "object",
                "description": "Rename image request",
                "required": [
                    "new_filename"
                ],
                "properties": {
                    "new_filename": {
                        "type": "string",
                        "example": "new_image_name.jpg"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer_auth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}